+++ includes/class-polygon-tweaks-textdomain.php
@@ -6,6 +6,8 @@
  * @package Polygon_Tweaks
  */
 
+defined( 'ABSPATH' ) || exit;
+
 
 
 
@@ -21,26 +23,27 @@
 class Polygon_Tweaks_Textdomain {
 
 	/**
-	 * Load plugin text-domain.
-	 *
-	 * Load the plugin text-domain and define the location of our translation files.
-	 * See examples below:
-	 *
-	 * - Global /languages/ folder: wp-content/languages/plugins/polygon-tweaks-en_US.mo
-	 * - Local /languages/ folder:  wp-content/plugins/polygon-tweaks/languages/polygon-tweaks-en_US.mo
-	 *
-	 * If no files are found in the global languages folder the plugin uses the files available in the
-	 * local folder.
+	 * Hook into actions and filters.
 	 *
 	 * @since 1.0.0
 	 */
-	public function load_plugin_textdomain() {
-		$locale = apply_filters( 'locale', get_locale(), POLYGON_TWEAKS_NAME );
+	public function init() {
+		add_action( 'plugins_loaded', array( $this, 'load_textdomain' ) );
+	}
 
-		// Load translation files from the global /languages/ folder.
-		load_textdomain( POLYGON_TWEAKS_NAME, trailingslashit( WP_LANG_DIR ) . 'plugins/' . POLYGON_TWEAKS_NAME . '-' . $locale . '.mo' );
 
-		// Load translation files from the local /languages/ folder.
-		load_plugin_textdomain( POLYGON_TWEAKS_NAME, false, plugin_basename( POLYGON_TWEAKS_DIR_PATH ) . '/languages/' );
+
+
+
+	/**
+	 * Load the plugin textdomain.
+	 *
+	 * The plugin tries to load the files from the global /languages/ folder first.
+	 * If it can't find any, it will load the files from the local /languages/ folder.
+	 *
+	 * @since 1.0.0
+	 */
+	public function load_textdomain() {
+		load_plugin_textdomain( 'polygon-tweaks', false, plugin_basename( POLYGON_TWEAKS_DIR_PATH ) . '/languages/' );
 	}
 }