+++ polygon-tweaks.php
@@ -5,7 +5,7 @@
  * Author:            Polygon Themes
  * Author URI:        https://polygonthemes.com
  * Description:       A simple plugin with tweaks and fixes for a better experience on your WordPress website.
- * Version:           1.1.6
+ * Version:           1.2.0
  * Requires PHP:      7.2
  * Requires at least: 5.0
  *
@@ -28,38 +28,27 @@
  * @package Polygon_Tweaks
  */
 
-
-
-
-
-/**
- * Abort if this file is called directly.
- */
-if ( ! defined( 'WPINC' ) ) {
-	die;
-}
+defined( 'ABSPATH' ) || exit;
 
 
 
 
 
-/**
- * Define plugin constants.
- */
-define( 'POLYGON_TWEAKS_VERSION', '1.1.6' );                         // Current plugin version.
-define( 'POLYGON_TWEAKS_NAME', 'polygon-tweaks' );                   // Unique plugin identifier.
+define( 'POLYGON_TWEAKS_VERSION', '1.2.0' );
+define( 'POLYGON_TWEAKS_SLUG', 'polygon-tweaks' );
 
-define( 'POLYGON_TWEAKS_MAIN_FILE', __FILE__ );                      // Path to main plugin file.
-define( 'POLYGON_TWEAKS_DIR_PATH', plugin_dir_path( __FILE__ ) );    // Path to plugin directory.
-define( 'POLYGON_TWEAKS_DIR_URL', plugin_dir_url( __FILE__ ) );      // URL to plugin directory.
+define( 'POLYGON_TWEAKS_FILE', __FILE__ );
+define( 'POLYGON_TWEAKS_DIR_URL', plugin_dir_url( __FILE__ ) );
+define( 'POLYGON_TWEAKS_DIR_PATH', plugin_dir_path( __FILE__ ) );
+
+define( 'POLYGON_TWEAKS_MIN_PHP_VERSION', '7.2' );
+define( 'POLYGON_TWEAKS_REC_PHP_VERSION', '8.0' );
 
 
 
 
 
 /**
- * Activate Polygon Tweaks.
- *
  * Code that runs during the plugin activation.
  *
  * @since 1.0.0
@@ -69,15 +58,13 @@ function activate_polygon_tweaks( $network_wide ) {
 	require_once POLYGON_TWEAKS_DIR_PATH . 'includes/class-polygon-tweaks-activator.php';
 	Polygon_Tweaks_Activator::activate( $network_wide );
 }
-register_activation_hook( POLYGON_TWEAKS_MAIN_FILE, 'activate_polygon_tweaks' );
+register_activation_hook( POLYGON_TWEAKS_FILE, 'activate_polygon_tweaks' );
 
 
 
 
 
 /**
- * Deactivate Polygon Tweaks.
- *
  * Code that runs during the plugin deactivation.
  *
  * @since 1.0.0
@@ -87,7 +74,7 @@ function deactivate_polygon_tweaks( $network_wide ) {
 	require_once POLYGON_TWEAKS_DIR_PATH . 'includes/class-polygon-tweaks-deactivator.php';
 	Polygon_Tweaks_Deactivator::deactivate( $network_wide );
 }
-register_deactivation_hook( POLYGON_TWEAKS_MAIN_FILE, 'deactivate_polygon_tweaks' );
+register_deactivation_hook( POLYGON_TWEAKS_FILE, 'deactivate_polygon_tweaks' );
 
 
 
@@ -96,18 +83,22 @@ register_deactivation_hook( POLYGON_TWEAKS_MAIN_FILE, 'deactivate_polygon_tweaks
 
 
 /**
- * Run Polygon Tweaks.
- *
- * Load and execute the code of our plugin if all requirements are met.
+ * Load and execute if all requirements are met.
  *
  * @since 1.0.0
  */
 function run_polygon_tweaks() {
+	require_once POLYGON_TWEAKS_DIR_PATH . 'includes/class-polygon-tweaks.php';
+	require_once POLYGON_TWEAKS_DIR_PATH . 'includes/class-polygon-tweaks-textdomain.php';
 	require_once POLYGON_TWEAKS_DIR_PATH . 'includes/class-polygon-tweaks-requirements.php';
+
+	$textdomain = new Polygon_Tweaks_Textdomain();
+	$textdomain->init();
+
 	$requirements = new Polygon_Tweaks_Requirements();
+	$requirements->init();
 
 	if ( $requirements->check() ) {
-		require_once POLYGON_TWEAKS_DIR_PATH . 'includes/class-polygon-tweaks.php';
 		$plugin = new Polygon_Tweaks();
 		$plugin->run();
 	}