+++ uninstall.php
@@ -6,16 +6,7 @@
  * @package Polygon_Tweaks
  */
 
-
-
-
-
-/**
- * Prevent unauthorized access.
- */
-if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
-	exit;
-}
+defined( 'WP_UNINSTALL_PLUGIN' ) || exit;